pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = ""
authors = ["anna-dm <anna-anna-frolova@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
